WebSocket servicve reorg
